Allow kqueue to emit directory change notifications when the contents changes.
As kqueue unfortunately doesn't have any mechanism for monitoring changes for
the contents of a directory (like inotify, and Windows), we must do this
ourselves explicitly.

We do this by watching the entire contents of the directory, thus, when a change
happens to a file, we'll emit a signal.

Also add some code to mangle RLIMIT_NOFILE, as at least on OS X, the default is
ridiculously low (256), which is not even really that good for the _old_ uses of
QFileSystemWatcher, let alone this functionality.